fix(github-reporter): Handle report edge cases
Let positional JSON report arguments override environment defaults so the documented CLI form works. Also keep Check Run summaries within GitHub's limit and carry rounded duration seconds into minutes.
