+++ packages/github-reporter/src/cli.ts
@@ -1,5 +1,6 @@
 #!/usr/bin/env node
 import { appendFile, readFile } from "node:fs/promises";
+import { parseCliArgs } from "./cli-options";
 import { collectEvalReport } from "./collect";
 import { publishCheckRun } from "./github";
 import { renderWorkflowCommands } from "./annotations";
@@ -7,30 +8,17 @@ import { renderJobSummary } from "./summary";
 import type { VitestJsonReport } from "./types";
 import { escapeCommandData } from "./utils";
 
-type CliOptions = {
-  jsonPath?: string;
-  summaryPath?: string;
-  summaryEnabled: boolean;
-  annotations: boolean;
-  checkRun: boolean;
-  failOnCheckError: boolean;
-  maxAnnotations?: number;
-  maxFailures?: number;
-  checkRunId?: number;
-  checkName?: string;
-  token?: string;
-  repository?: string;
-  sha?: string;
-  workspace?: string;
-};
-
 main().catch((error) => {
   console.error(error instanceof Error ? error.message : String(error));
   process.exitCode = 1;
 });
 
 async function main() {
-  const options = parseArgs(process.argv.slice(2));
+  const options = parseCliArgs(process.argv.slice(2));
+  if (options.help) {
+    console.log(usage());
+    return;
+  }
   if (!options.jsonPath) {
     console.error(usage());
     process.exitCode = 1;
@@ -89,99 +77,6 @@ async function main() {
   }
 }
 
-function parseArgs(args: string[]): CliOptions {
-  const options: CliOptions = {
-    jsonPath: process.env.VITEST_EVALS_JSON_REPORT ?? "vitest-results.json",
-    summaryPath: process.env.GITHUB_STEP_SUMMARY,
-    summaryEnabled: true,
-    annotations: process.env.GITHUB_ACTIONS === "true",
-    checkRun: false,
-    failOnCheckError: false,
-  };
-
-  for (let index = 0; index < args.length; index += 1) {
-    const arg = args[index];
-    switch (arg) {
-      case "--json":
-        options.jsonPath = readValue(args, ++index, arg);
-        break;
-      case "--summary":
-        options.summaryPath = readValue(args, ++index, arg);
-        options.summaryEnabled = true;
-        break;
-      case "--no-summary":
-        options.summaryEnabled = false;
-        break;
-      case "--annotations":
-        options.annotations = true;
-        break;
-      case "--no-annotations":
-        options.annotations = false;
-        break;
-      case "--check-run":
-        options.checkRun = true;
-        break;
-      case "--fail-on-check-error":
-        options.failOnCheckError = true;
-        break;
-      case "--max-annotations":
-        options.maxAnnotations = readInteger(args, ++index, arg);
-        break;
-      case "--max-failures":
-        options.maxFailures = readInteger(args, ++index, arg);
-        break;
-      case "--check-run-id":
-        options.checkRunId = readInteger(args, ++index, arg);
-        options.checkRun = true;
-        break;
-      case "--check-name":
-        options.checkName = readValue(args, ++index, arg);
-        break;
-      case "--token":
-        options.token = readValue(args, ++index, arg);
-        break;
-      case "--repo":
-        options.repository = readValue(args, ++index, arg);
-        break;
-      case "--sha":
-        options.sha = readValue(args, ++index, arg);
-        break;
-      case "--workspace":
-        options.workspace = readValue(args, ++index, arg);
-        break;
-      case "--help":
-      case "-h":
-        console.log(usage());
-        process.exit(0);
-        return options;
-      default:
-        if (!arg.startsWith("-") && !options.jsonPath) {
-          options.jsonPath = arg;
-          break;
-        }
-        throw new Error(`Unknown argument: ${arg}`);
-    }
-  }
-
-  return options;
-}
-
-function readValue(args: string[], index: number, flag: string) {
-  const value = args[index];
-  if (!value) {
-    throw new Error(`Missing value for ${flag}`);
-  }
-  return value;
-}
-
-function readInteger(args: string[], index: number, flag: string) {
-  const value = Number.parseInt(readValue(args, index, flag), 10);
-  if (!Number.isFinite(value) || value < 0) {
-    throw new Error(`Invalid integer for ${flag}`);
-  }
-  return value;
-}
-
 function warn(message: string) {
   if (process.env.GITHUB_ACTIONS === "true") {
     console.error(`::warning::${escapeCommandData(message)}`);
@@ -192,7 +87,7 @@ function warn(message: string) {
 
 function usage() {
   return [
-    "Usage: vitest-evals-github-report [--json <vitest-results.json>]",
+    "Usage: vitest-evals-github-report [vitest-results.json] [--json <path>]",
     "",
     "Options:",
     "  --summary <path>          Write job summary markdown to this path",
